Clean up some code
Former-commit-id: 01788d37c2da84116c14d43ca318e5be4134d86f [formerly 5a4bd2b503f8ace2de3199e446123d2ebfebcc5e] [formerly e08e7b4831f4221b8a424d7a023e6e31081c2f63 [formerly 412302cba9aeeb89cb2d4b31a7b711c789ad8444]]
Former-commit-id: 09c0406c65994e0f153eb472cc715839eef33ef4 [formerly fd5a153e97d8dc353b744367571b97396eec2cf8]
Former-commit-id: ee4ef949183459a4067c5cbe2290c555c3753617

+++ src/Nazara/Network/Win32/IpAddressImpl.cpp
@@ -56,7 +56,7 @@ namespace Nz
 			std::array<char, NI_MAXHOST> hostnameBuffer;
 			std::array<char, NI_MAXSERV> serviceBuffer;
 
-			int result = getnameinfo(socketAddress, socketLen, hostnameBuffer.data(), static_cast<std::size_t>(hostnameBuffer.size()), serviceBuffer.data(), static_cast<std::size_t>(serviceBuffer.size()), flags);
+			int result = getnameinfo(socketAddress, socketLen, hostnameBuffer.data(), static_cast<DWORD>(hostnameBuffer.size()), serviceBuffer.data(), static_cast<DWORD>(serviceBuffer.size()), flags);
 			if (result == 0)
 			{
 				if (hostname)
